requirements.txt
appier==1.5.28
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.9
defusedxml==0.4.1
Django==1.9.6
django-bootstrap3==7.0.1
django-crontab==0.7.1
django-extensions==1.6.7
facebook-sdk==1.0.0
httplib2==0.9.2
ipython==4.2.0
ipython-genutils==0.1.0
oauth2==1.9.0.post1
oauthlib==1.1.2
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
PyJWT==1.4.0
python-dateutil==2.5.3
python-openid==2.2.5
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.4
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
